refactor(models): colocate sparse MLA operators
